+++ packages/icon-font/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## v2.0.0 (Unreleased)
+
+### Changed
+
+-   Changed import path of icon font files to `icons`.
+
+#### React
+
+```diff
+- import '@brightlayer-ui/icons/iconfont/BrightlayerUIIcons.css'
++ import '@brightlayer-ui/icons/BrightlayerUIIcons.css'
+```
+
+#### Angular
+
+```diff
+- "styles": ["src/styles.scss", "./node_modules/@brightlayer-ui/icons/iconfont/BrightlayerUIIcons.css"]
++ "styles": ["src/styles.scss", "./node_modules/@brightlayer-ui/icons/BrightlayerUIIcons.css"]
+```
+
 ## v1.11.0 (Aug 10, 2022)
 
 ### Added